requirements/base.txt
asgiref==3.3.4
Django==3.2.3
django-comment-migrate==0.1.2
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework_simplejwt==4.7.1
django-tenants==3.3.2
#mysqlclient==2.0.3
#pymysql==1.0.2
PyJWT==2.1.0
psycopg2-binary==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
user-agents==2.2.0
drf-yasg==1.20.0
daphne==3.0.2
django-simple-captcha==0.5.14
gitpython==3.1.20
pypinyin==0.42.0
whitenoise==5.3.0